+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ##### Enhancements
 
+* Add accessors for working with Environment Variables in xcscheme files
+  [Justin Martin](https://github.com/justinseanmartin)
+  [Xcodeproj#326](https://github.com/CocoaPods/Xcodeproj/pull/326)
+
 * Add method to create new variant groups (groups for localized versions of the same file)  
   [Tim Bodeit](https://github.com/timbodeit)
   [Xcodeproj#315](https://github.com/CocoaPods/Xcodeproj/pull/315)